SovereignAgents/invoice_processor/requirements.txt
fastapi
uvicorn[standard]
gradio
PyMuPDF
pytesseract
Pillow
pydantic
requests
python-dotenv
